Calendar invite email improvements
